edgeManager/backend/requirements.txt
Flask==2.2.2
flask-restx==1.0.3
flask-cors==3.0.10
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.6
kubernetes==23.6.0
elasticsearch==7.11.0
gunicorn==20.1.0
